requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
ftfy>=6.1.1
regex>=2023.6.3
tqdm>=4.65.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
wandb>=0.15.0
open_clip_torch>=2.20.0
einops>=0.6.1
timm>=0.9.2
